temp_precip_plugin/requirements.txt
affine==2.4.0
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
asciitree==0.3.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.22.2
botocore==1.29.76
camel-converter==3.0.0
certifi==2023.5.7
cffi==1.15.1
cfgrib==0.9.10.3
cftime==1.6.2
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
cycler==0.11.0
dask==2022.4.1
dataclasses-json==0.5.7
docopt==0.6.2
eccodes==1.5.2
entrypoints==0.4
fasteners==0.18
findlibs==0.0.5
Fiona==1.9.3
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
GDAL==3.5.1
geopandas==0.10.2
gitdb==4.0.11
GitPython==3.1.37
h5netcdf==1.0.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
locket==1.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.5.1
meilisearch==0.28.0
multidict==6.0.4
munch==2.5.0
mypy-extensions==1.0.0
netCDF4==1.5.8
numcodecs==0.11.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
papipyplug==2023.10.3
partd==1.4.0
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.5.0
pooch==1.7.0
pycparser==2.21
pydantic==1.10.8
pydap==3.4.0
# pydsstools==2.2
pyparsing==3.0.9
pyproj==3.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3
PyYAML==6.0
rasterio==1.3.6
rasterstats==0.18.0
requests==2.30.0
rioxarray==0.11.1
Rtree==1.0.0
s3fs==2023.5.0
s3transfer==0.5.2
scikit-learn==1.2.2
scipy==1.10.1
shapely==2.0.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
soupsieve==2.4.1
threadpoolctl==3.1.0
toolz==0.12.0
typing-extensions==4.5.0
typing-inspect==0.8.0
tzdata==2023.3
urllib3==1.26.15
WebOb==1.8.7
wrapt==1.15.0
xarray==2022.3.0
yarl==1.9.2
zarr==2.14.2
zipp==3.15.0
